defaultIndex: return 404 + retry index logic
This extends the defaultIndex logic to return a 404 for the index.html if it is not found.
Additionally, we keep looking for the "defaultIndex.html" file in the URL path, by splitting off the last "/" and keep going up until we find it, or we are at the root (and need to return a 404).